+++ src/core/navigation/tab/index.tsx
@@ -1,15 +1,16 @@
 import React from 'react'
+import type { ReactNode } from 'react'
 import MuiTab from '@mui/material/Tab'
 import type { DefaultProps } from '../../types'
 import type { TabOwnProps } from '@mui/material/Tab'
 
 interface TabProps extends Omit<DefaultProps, 'children'>, TabOwnProps {
     href?: string
-    label?: string
     value?: unknown
     minWidth?: string
     disabled?: boolean
     disableRipple?: boolean
+    label?: string | ReactNode
     icon?: string | JSX.Element
 }